+++ apps/remnant2toolkit/app/(actions)/builds/get-community-builds.ts
@@ -74,10 +74,13 @@ export async function getCommunityBuilds({
     releases,
     relic,
     patchAffected,
-    withQuality,
+    // withQuality,
     withVideo,
     withReference,
   } = buildListFilters
+
+  // TODO Quality Builds
+  const withQuality = false
 
   if (releases.length === 0) return { items: [], totalItemCount: 0 }
 
@@ -121,7 +124,6 @@ export async function getCommunityBuilds({
       orderBySegment,
       whereConditions,
       searchText: trimmedSearchText,
-      limitToQualityBuilds: withQuality,
     }),
     communityBuildsCountQuery({
       whereConditions,