+++ apps/remnant2toolkit/app/(components)/filters/builds/build-misc-filter.tsx
@@ -16,10 +16,11 @@ export const MINIMUM_QUALITY_DESCRIPTION_LENGTH = 200
 
 export function BuildMiscFilter({ value, onChange }: Props) {
   const options = [
-    {
-      label: `Only Quality Builds`,
-      value: BUILD_FILTER_KEYS.WITHQUALITY,
-    },
+    // TODO Quality builds
+    // {
+    //   label: `Only Quality Builds`,
+    //   value: BUILD_FILTER_KEYS.WITHQUALITY,
+    // },
     {
       label: 'Only Builds w/ Video',
       value: BUILD_FILTER_KEYS.WITHVIDEO,