+++ apps/remnant2toolkit/app/profile/[userId]/favorited-builds/actions/get-favorite-builds.ts
@@ -75,10 +75,13 @@ export async function getFavoritedBuilds({
     searchText,
     releases,
     patchAffected,
-    withQuality,
+    // withQuality,
     withVideo,
     withReference,
   } = buildListFilters
+
+  // TODO Quality Builds
+  const withQuality = false
 
   if (releases.length === 0) return { items: [], totalItemCount: 0 }
 
@@ -124,7 +127,6 @@ ${limitToQualityBuilds(withQuality)}
       orderBySegment,
       whereConditions,
       searchText: trimmedSearchText,
-      limitToQualityBuilds: withQuality,
     }),
     communityBuildsCountQuery({
       whereConditions,