+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2024-07-18
+
+### Changed
+
+- Temporarily disabled the ability to filter by Quality Builds due to a technical issue with the database provider. Hoping to resolve this soon.
+
 ## 2024-07-15
 
 ### Fixed